Add files via upload
This system is for Students who feel they want to sell their products or buy.  the platform allows one to create account and login, access the products that are available. get the number displayed below the product and contact the seller. once it is sold, the seller deletes it from display